refactor(cpp): remove useless exception files
